Reduce seed packet size

+++ Scripts/Gameplay/CharacterSystems/CharacterActionsSystem/DefaultCharacterAttackComponent.cs
@@ -87,7 +87,7 @@ namespace MultiplayerARPG
             animSpeedRate *= Entity.GetAnimSpeedRate(Entity.AnimActionType);
 
             // Get simulate seed for simulation validating
-            int simulateSeed = Random.Range(int.MinValue, int.MaxValue);
+            byte simulateSeed = (byte)Random.Range(byte.MinValue, byte.MaxValue);
             simulatingHitIndex = 0;
             try
             {
